Ideas to improve test maintenance
1) Seperate into specs to easier to run
2) Improve names
3) Split dependencies into named It()s rather then just part of test
eg it('Creates Instance Template For Purpose Of Editing in Next Step',)
